changed to double usage, added mini map and towerselection

+++ JTD/src/jtd/def/AnimatedEntityDef.java
@@ -10,17 +10,17 @@ package jtd.def;
  */
 public abstract class AnimatedEntityDef extends EntityDef{
 
-	public float[] times = {1f};
+	public double[] times = {1f};
 	public boolean isCyclic = true;
 
-	public void fillImage(String name, float size, float time){
+	public void fillImage(String name, double size, double time){
 		fillImage(name, size);
-		times = new float[]{time};
+		times = new double[]{time};
 	}
 	
-	public void fillImages(String nameStart, String nameEnd, int imageCount, int digitCount, float sizes, float times){
+	public void fillImages(String nameStart, String nameEnd, int imageCount, int digitCount, double sizes, double times){
 		fillImages(nameStart, nameEnd, imageCount, digitCount, sizes);
-		this.times = new float[imageCount];
+		this.times = new double[imageCount];
 		for(int i=0; i<imageCount; i++){
 			this.times[i] = times;
 		}
