+++ JTD/src/jtd/def/MobDef.java
@@ -14,17 +14,17 @@ import jtd.entities.ParticleFactory;
 public class MobDef extends AnimatedEntityDef{
 
 	public int size = 1;
-	public float maxHP = 1f;
-	public float hpRegen = 0f;
-	public float maxShield = 0f;
-	public float shieldRegen = 0f;
-	public float armor = 0f;
-	public float speed = 1f;
-	public float radius = 0.5f;
+	public double maxHP = 1f;
+	public double hpRegen = 0f;
+	public double maxShield = 0f;
+	public double shieldRegen = 0f;
+	public double armor = 0f;
+	public double speed = 1f;
+	public double radius = 0.5f;
 	public ParticleFactory[] hitParticleFacts = {};
 	public ParticleFactory[] deathParticleFacts = {};
-	public float[] damagePerParticle = {};
+	public double[] damagePerParticle = {};
 	public int[] deathParticleCounts = {};
 
-	public void defTick(Mob mob, float time){}
+	public void defTick(Mob mob, double time){}
 }
